Update security groups on device_owner changes
During Mitaka cycle, Neutron changed the way to update device_owner
for router-interface-add via port. [1]  Unfortunately it broke our
translator's assumption of the sequence of events.  This commit
makes the translator deal with the new way, that is, the device_owner
change comes as a normal update_port request before add_router_interface
request.  Note that this commit doesn't remove the old way, as it's
still necessary for Liberty.

This commit also fixes SGs for VIP ports.  Neutron ports with
device_owner starting with "network:" are considered "trusted".
The reference implmentation ignores SGs and port_security_enabled
values on those ports, as they had port_security_enabled=False.
VIP ports, whose device_owner is "neutron:xxx", should not be
considered trusted.

[1] If0178887282456842b6078a851a9233cb58a391a

Fixes MI-1185
Fixes MI-1186

+++ midonet-cluster/src/main/scala/org/midonet/cluster/services/c3po/translators/PortTranslator.scala
@@ -94,16 +94,17 @@ class PortTranslator(protected val storage: ReadOnlyStorage,
 
         val portId = nPort.getId
         val portContext = initPortContext
-        if (isVifPort(nPort)) {
+        if (!isTrustedPort(nPort)) {
             // Generate in/outbound chain IDs from Port ID.
             midoPortBldr.setInboundFilterId(inChainId(portId))
             midoPortBldr.setOutboundFilterId(outChainId(portId))
-
+        }
+        updateSecurityBindings(nPort, None, midoPortBldr, portContext)
+        if (isVifPort(nPort)) {
             // Add new DHCP host entries
             updateDhcpEntries(nPort,
                               portContext.midoDhcps,
                               addDhcpHost)
-            updateSecurityBindings(nPort, None, midoPortBldr, portContext)
         } else if (isDhcpPort(nPort)) {
             updateDhcpEntries(nPort,
                               portContext.midoDhcps,
@@ -175,14 +176,16 @@ class PortTranslator(protected val storage: ReadOnlyStorage,
         }
 
         val portContext = initPortContext
+        if (!isTrustedPort(nPort)) {
+            portContext.chains ++= deleteSecurityChainsOps(nPort.getId)
+            portContext.updatedIpAddrGrps ++=
+                removeIpsFromIpAddrGroupsOps(nPort)
+        }
         if (isVifPort(nPort)) { // It's a VIF port.
             // Delete old DHCP host entries
             updateDhcpEntries(nPort,
                               portContext.midoDhcps,
                               delDhcpHost)
-            portContext.chains ++= deleteSecurityChainsOps(nPort.getId)
-            portContext.updatedIpAddrGrps ++=
-                removeIpsFromIpAddrGroupsOps(nPort)
             // Delete the ARP entries for associated Floating IP.
             midoOps ++= deleteFloatingIpArpEntries(nPort)
         } else if (isDhcpPort(nPort)) {
@@ -230,6 +233,15 @@ class PortTranslator(protected val storage: ReadOnlyStorage,
         if (portChanged(nPort, oldNPort)) {
             val bldr = mPort.toBuilder.setAdminStateUp(nPort.getAdminStateUp)
 
+            if (isTrustedPort(nPort)) {
+                bldr.clearInboundFilterId()
+                bldr.clearOutboundFilterId()
+            } else {
+                // Generate in/outbound chain IDs from Port ID.
+                bldr.setInboundFilterId(inChainId(portId))
+                bldr.setOutboundFilterId(outChainId(portId))
+            }
+
             // Neutron may specify binding information for router interface
             // ports on edge routers. For VIF/DHCP ports, binding information
             // is controlled by mm-ctl, and we shouldn't change it.
@@ -274,8 +286,10 @@ class PortTranslator(protected val storage: ReadOnlyStorage,
             }
         }
 
+        val portContext = initPortContext
+        updateSecurityBindings(nPort, Some(oldNPort), mPort, portContext)
+
         if (isVifPort(nPort)) { // It's a VIF port.
-            val portContext = initPortContext
             // Delete old DHCP host entries
             updateDhcpEntries(oldNPort,
                               portContext.midoDhcps,
@@ -284,13 +298,11 @@ class PortTranslator(protected val storage: ReadOnlyStorage,
             updateDhcpEntries(nPort,
                               portContext.midoDhcps,
                               addDhcpHost)
-            updateSecurityBindings(nPort, Some(oldNPort), mPort, portContext)
-            addMidoOps(portContext, midoOps)
-
             // TODO: Can other port types update MAC/IP?
             midoOps ++= macTableUpdateOps(oldNPort, nPort)
             midoOps ++= arpTableUpdateOps(oldNPort, nPort)
         }
+        addMidoOps(portContext, midoOps)
 
         // TODO if a DHCP port, assert that the fixed IPs haven't changed.
 
@@ -336,6 +348,7 @@ class PortTranslator(protected val storage: ReadOnlyStorage,
     private def portChanged(newPort: NeutronPort,
                             curPort: NeutronPort): Boolean = {
         if (newPort.getAdminStateUp != curPort.getAdminStateUp ||
+            isTrustedPort(newPort) != isTrustedPort(curPort) ||
             hasBinding(newPort) != hasBinding(curPort)) return true
 
         hasBinding(newPort) &&
@@ -458,12 +471,27 @@ class PortTranslator(protected val storage: ReadOnlyStorage,
     private def updatePortSgAssociations(nPort: NeutronPort,
                                          nPortOld: Option[NeutronPort],
                                          portCtx: PortContext): Unit = {
-        val newIps = nPort.getFixedIpsList.asScala map (_.getIpAddress)
-        val newSgs = nPort.getSecurityGroupsList.asScala
+        def getFixedIps(nPort: NeutronPort) = {
+            if (isTrustedPort(nPort)) {
+                Seq()
+            } else {
+                nPort.getFixedIpsList.asScala.map(_.getIpAddress)
+            }
+        }
 
-        val oldIps = nPortOld.toList.flatMap(
-            _.getFixedIpsList.asScala.map(_.getIpAddress))
-        val oldSgs = nPortOld.toList.flatMap(_.getSecurityGroupsList.asScala)
+        def getSecurityGroups(nPort: NeutronPort) = {
+            if (isTrustedPort(nPort)) {
+                Seq()
+            } else {
+                nPort.getSecurityGroupsList.asScala
+            }
+        }
+
+        val newIps = getFixedIps(nPort)
+        val newSgs = getSecurityGroups(nPort)
+
+        val oldIps = nPortOld.toList.flatMap(getFixedIps)
+        val oldSgs = nPortOld.toList.flatMap(getSecurityGroups)
 
         val removedSgs = oldSgs diff newSgs
         val addedSgs = newSgs diff oldSgs
@@ -533,33 +561,36 @@ class PortTranslator(protected val storage: ReadOnlyStorage,
         val outChainId = mPort.getOutboundFilterId
         val antiSpfChainId = antiSpoofChainId(portId)
 
-        // Create in/outbound chains.
-        val inChain = newChain(inChainId, egressChainName(portId))
-        val outChain = newChain(outChainId, ingressChainName(portId))
-        val antiSpoofChain = newChain(
-            antiSpfChainId, antiSpoofChainName(portId),
-            jumpRuleIds = Seq(antiSpoofChainJumpRuleId(portId)))
-
         // If this is an update, clear the chains. Ideally we would be a bit
         // smarter about this and only delete rules that actually need deleting,
         // but this is simpler.
-        nPortOld match {
-            case Some(_) =>
+        if (nPortOld.exists(!isTrustedPort(_))) {
+            if (isTrustedPort(nPort)) {
+                portCtx.chains ++= deleteSecurityChainsOps(portId)
+            } else {
                 portCtx.inRules ++= deleteRulesOps(
                     storage.get(classOf[Chain], inChainId).await())
                 portCtx.outRules ++= deleteRulesOps(
                     storage.get(classOf[Chain], outChainId).await())
                 portCtx.antiSpoofRules ++= deleteRulesOps(
                     storage.get(classOf[Chain], antiSpfChainId).await())
-            case None =>
-                portCtx.chains += (
-                    Create(inChain), Create(outChain), Create(antiSpoofChain))
+            }
+        } else if (!isTrustedPort(nPort)) {
+            // Create in/outbound chains.
+            val inChain = newChain(inChainId, egressChainName(portId))
+            val outChain = newChain(outChainId, ingressChainName(portId))
+            val antiSpoofChain = newChain(
+                antiSpfChainId, antiSpoofChainName(portId),
+                jumpRuleIds = Seq(antiSpoofChainJumpRuleId(portId)))
+
+            portCtx.chains += (
+                Create(inChain), Create(outChain), Create(antiSpoofChain))
         }
 
         updatePortSgAssociations(nPort, nPortOld, portCtx)
 
         // Add new rules if appropriate.
-        if (nPort.getPortSecurityEnabled) {
+        if (!isTrustedPort(nPort) && nPort.getPortSecurityEnabled) {
 
             buildAntiSpoofRules(portCtx, nPort, portId, inChainId)
